Update DualSVM class with training data and labels
The constructor of `DualSVM` class has been updated to include `training_data` and `training_labels` as parameters. The `dimension` attribute is now defined based on the size of the training labels. The `_polynomial_kernel` method no longer needs the `degree` parameter since it can now access the `degree` attribute directly. The `fit` and `predict` methods have been implemented with the necessary logic.